requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
greenlet==3.0.1
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.5
MarkupSafe==3.0.2
npm==0.1.1
optional-django==0.1.0
packaging==24.2
playwright==1.40.0
pluggy==1.5.0
pyee==11.0.1
pytest==8.3.5
pytest-base-url==2.1.0
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-playwright==0.7.0
python-slugify==8.0.4
requests==2.32.3
text-unidecode==1.3
typing_extensions==4.12.2
undetected-playwright-patch==1.40.0.post1700587210000
urllib3==2.3.0
